more personal elements added, GUI is also improved
