+++ src/test/java/Chatbot/alain/UiTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 
+import chatbot.alain.uis.Ui;
 import org.junit.jupiter.api.Test;
 
 public class UiTest {
